# Use faster urlparse alternative