Hi ! This improves the middleware performance by roughly 1-5% (depending on the use case) by replacing `urlib.parse.urlparse` usage by the almost equivalent `urlib.parse.urlsplit` which perform 2x faster. 

```
In [5]: %timeit urlparse("https://docs.djangoproject.com/en/4.1/;deprecated=true?q=1&e=2#123")
1.27 µs ± 5.06 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [6]: %timeit urlsplit("https://docs.djangoproject.com/en/4.1/;deprecated=true?q=1&e=2#123")
671 ns ± 4.21 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
This is really a slight improvement but since middleware are processed every request, small performance improvements adds up.